requirements.txt
torch
torchvision
pyyaml
Pillow
numpy
opencv-python
easydict
matplotlib
tqdm